requirements.txt
pandas>=2.1.0
numpy>=1.24.0
twelvedata>=1.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
MetaTrader5>=5.0.45
ta>=0.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
